move Club entity into root model package; update imports
