+++ dashboard-server/src/main/java/com/footballstatsdashboard/api/model/Club.java
@@ -1,4 +1,4 @@
-package com.footballstatsdashboard.api.model.club;
+package com.footballstatsdashboard.api.model;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonInclude;
