+++ dashboard-server/src/main/java/com/footballstatsdashboard/resources/ClubResource.java
@@ -1,7 +1,7 @@
 package com.footballstatsdashboard.resources;
 
 import com.footballstatsdashboard.api.model.User;
-import com.footballstatsdashboard.api.model.club.Club;
+import com.footballstatsdashboard.api.model.Club;
 import com.footballstatsdashboard.api.model.club.ClubSummary;
 import com.footballstatsdashboard.api.model.club.SquadPlayer;
 import com.footballstatsdashboard.services.ClubService;
