+++ dashboard-server/src/test/java/com/footballstatsdashboard/services/ClubServiceTest.java
@@ -1,7 +1,7 @@
 package com.footballstatsdashboard.services;
 
 import com.footballstatsdashboard.ClubDataProvider;
-import com.footballstatsdashboard.api.model.club.Club;
+import com.footballstatsdashboard.api.model.Club;
 import com.footballstatsdashboard.api.model.club.ClubSummary;
 import com.footballstatsdashboard.api.model.club.ImmutableSquadPlayer;
 import com.footballstatsdashboard.api.model.club.SquadPlayer;
